Add login-fire-button element
This commit adds a new element called "login-fire-button". It is a
button that offers an easy way to authenticate a user using the Firebase
Auth API. This element uses the "login-fire-common-behavior" that can be
implemented by other Polymer elements. This commit contains also a
collection of SVG icons for the Google, Twitter, Facebook and GitHub
logos. There is also a demo page.