Use povider to compute button text
This commit adds the provider as argument of the function to compute the
text of the button. It is for fixing a race issue with this computed
property and the config inside the behavior. In fact it's normal. To
define the button text we need to know if the user is logged in and the
provider that is used by the button.