Add signedId and statusKnwon to the behavior
This commit copies the properties "signedIn" and "statusKnown" from the
firebase-auth element to the LoginFireCommonBehavior. They were present
in the original source code but even I don't think those properties are
relevant for the login-fire element. I mean, they are accessible through
the "auth" element inside the login-fire-button that is public.