Refactor email-login-fire
The element email-login-fire was renamed login-fire-form. The purpose of
the refactoring is to uniformize the collection by using the
LoginFireCommonBehavior. Using the behavior, the elements can be linked
between each others and used into a composition to be automatically
synchronized.

Again, the goal is to centralize the interaction with Firebase Auth API,
make maintanability easier and as far as possible, provide to the
community a behavior plug-and-play to make an element "login-fire-like".