Add login-fire-social-behavior
This commit moves the sign-in process with social provider
authentication out of the LoginFireCommonBehavior behavior
put it into a new LoginFireSocialBehavior behavior. By
doing this, the common behavior is only related to the
Firebase Auth API and the properties shared between all
elements that are implementing the LoginFireCommoBehavior
or one of the extending behaviors.