Make sign out feature optional
The sign out feature is now optional. By default, the elements
implementing the LoginFireCommonBehavior behavior cannot ask for signing
out. For example, with this commit, the element `login-fire-button` does
not change of state anymore and can't log out a user. The feature can be
enabled by setting its attribute "canSignOut" to true.