Updating buttons style in login-social
Before this commit, when the buttons text was too wide for the default
size (210px), it was splitted in two lines. In happens when the text
is translated in French for example. So I changed a little the default
behavior like this:

* a button alone takes space as much as it needs according to its text
* all login-social buttons have the same size which is the login-social
width (default is 210px)
* when the login-social displays buttons vertically, all buttons has its
own widht (related to its text)
* widths styles still can be updated using CSS custom variables