fix: version-specific code
The `login-fire-common-behavior` behavior has a callback function that
requires Polymer version-specific code. With Polymer 1, the callback is
(probably) executed without considering the "readOnly" aspect of the
property "resources". I mean, the setter for this property is not yet
generated. On the contrary, Polymer 2 seems respect this specificity
of the property so we can/have to use the special setter.