Review login-form structure
First of all, the source code indentation was reviewed for homogeneity.

Secondly, the `login-fire-form` was restructured. The `form` tag is gone
because we are using the Firebase API and the form is never submitted.
Moreover, there was two inputs for the email address for sign-up and
sign-in, now there is only one.