fix: login-fire-form width
The default width of the `login-fire-form` element was fixed to the
same value of the `login-fire-social` element.

Before this commit, when an error/info message was displayed on the
`login-fire-form` element, this one was growing according to the
space required by the message to be fully displayed on one line.
It's due to the fact the form container is a layout. Now the
element width doe not change.

A `wide` attribute was added to allow the form filled its parent.
For more customization, please use standard CSS rules.

Fixes #115