Refactor login-fire-social-behavior
The `login-fire-social-behavior` is only used by the
`login-fire-button`. Before this commit, it had 3 properties for
defining scopes to be set to the OAuth social provider. The fact is a
button can use only one provider at the time and should have scopes only
for one OAuth provider.

With this commit, the `login-fire-social-behavior` has only one property
`scopes` associated to its OAuth social provider. Moreover, the
providers dictionary was updated to avoid conditional statements.

BREAKING CHANGE:
The provider "password" was renamed "email" for matching purpose, the
Firebase's provider being "EmailAuthProvider".