fix(style): divider word properly aligned
Before this commit, for drawing a stripe to split the panel in 2 parts,
we were using border of a div tag. It introduces issue for aligning
the word "or" displayed on the middle of this one.

Now we are using background color that is easier to see the
representation in (our?) my mind. Because the stripe thickness is 1px,
if we want to have the same number of pixels on each sides the line
height of the word has to be odd. The original height was 14px so I've
simply increased by 1px to not modify too much the display.

fixes #123