chore(demo): update login-fire-button demo page
Before this commit, the demo page of the login-fire-button element was
not displaying the code snippet. By adding the `template` tags around
the `dom-bind` ones, the demo AND the code snippet are visible.

A little bit style was also added to make the page prettier. The titles
have now the same color of the Convoo logo, a little padding to make
them more visible and the family-font is Roboto.

The code was also formatted to make the lines easier to reader and
compare. This set of work will be applied on all demo pages.