fix(email-login-fire): non-unique input ids solved
After a Chrome update and before this commit, the email-login-fire
element was throwing a "non-unique input ids" error. It's due to the fact
the element is using Polymer v1 that does not fully encapsulate elements
inside the shadow root. This behavior is normal. Polymer v1 is using
Shady DOM instead of Shadow DOM.

As you can expect, the issue was not encountered with the element using
Polymer v2.

Fixes #143