perf(behaviors): increase control on sign-in/out processes
Firstly, the `login-fire-button` is now using the new added
`paper-social-button` element (by composition). The Firebase Auth
element was removed from its template. The fact is when the Firebase
Auth element is attached to the DOM it subscribes to Firebase and
tries to log the user in when possible. Because of it, `login-fire`
elements lose control on the sign-in/sign-out process progression and
can't inform the user in real time. To solve it, I introduced a new
property `auto`. When this property is `false` the Firebase Auth is
not created and not attached to the DOM without user action (form
validation, social button click). It introduces another minor side
effect which is `login-fire` elements that do not have the `auto`
property to `true` are not sync with those who have it till the user
has interaction with them.

Secondly, now the elements use the user representation stored by
Firebase Auth. It means the local representation is always prioritized
over the network. As you can imagine, you get instant connection.

Thirdly, the events firing system was also improved. Now it uses an
observer that ensure to fire events at the last time, after all others
properties are up to date. For example, it avoids to accidently ignore
a signIn/signOut call because the progress state is not yet updated or
get the previous user representation when we are listening on the
"signedin" and "signedout" events.

Finally, unit test suites were added. To be honest, without unit tests
and integration tests I'm not able to really know how much I have potentially
broke the elements, especially with the low frequency I work on this
project. I expect you have issue with rendering if you have fix some
locally by adding some CSS rules (because the apper-social-button fixes
some CSS issues by itself). And I guess you have automatic connection
that is not the default behavior here. We can discuss about the to turn
the default value of `auto` property to `true`. Until then, you have to
add `auto` attribute on all your login-fire elements.

BREAKING CHANGE:
The `signedin` event does not contain the user's credential anymore.
`login-fire` elements do not automatically try to log the user in
(see `auto` property).

Related: #165, #51, #149, #131