perf(login-fire-social): use paper-social-buton and add tests
Before this commit the `login-fire-social` element was using the `login-fire-button` element. This
one was embedded behaviors for i18n and authentication. So, when the login-fire-social was
displaying 5 buttons, the behaviors for localize texts and the behaviors for sign-in/up was imported
5 times. Now, using the `paper-social-button` (that does not include behavior at all) the
`login-fire-social` element is lighter, simpler to understand and maintain. This commit also adds
unit test suites.