perf(login-fire-form): add UT, strong password validation and refactor
First of all, a bunch of unit tests are added with this commit. It will be easier to maintain the
element in the future.

Now, the email field is checked before sending the form. The value must look like a real email
address. It's a simple check to ensure not request Firebase with random strings. The password
field is also checked before the form is sent. Now by default, the password must have 8 chars
and contains lowercase letters, uppercase letters, special char and digits. The regular
expression use for the validation can be changed (using the property `passwordRegex`).

Before this commit the form view was not reset after the form was sent. If the user sent the form
to sign-up successfully, when the user signed-out they see the sign-up view. Same issue with the
reset password view. Now, when the user signs out the form displays the "sign-in" view.

Finally, the element can throw 3 new events (which are "cancel", "confirm" and
"reset-password-email-sent").

BREAKING CHANGE: The property "showResetPW" was renamed "showResetPassword". Now and by default, the
email address format is checked and the password must be relatively strong else the form can't be
sent. The method "toggleShowResetPW" was renamed "toggleShowResetPassword".

fix #87, fix #122, re #51