fix(login-fire-button): ensure child has same width
Before this commit the `paper-social-button` width could not be the same
as its parent. So when user was setting the width of `login-fire-button`
the button was not applying it. Now, it does so you can change the width
of the button directly on the `login-fire-button` and not only on the
encapsulated `paper-social-button`.