refactor(login-fire): remove login-fire-common-behavior
Before this commit, to expose the same properties as its children, the
`login-fire` element was implementing the common behavior. I noticed
not all properties were properly downward/upward forwarded. Moreover,
children were able to overwrite the value set by another, depending on
the execution order (it was the case of the `inProgress` property).

The purpose of this commit is to turn the `login-fire` element into a
simple container (implementing the composition pattern).