fix(i18n): dynamically load locales
This commit replaces the static locales loading by a dynamic one that
allow the user to change the resources file after having attached the
element to the DOM. Sometimes you need to do some operation before
displaying the login page or update it. Now you can change the locales
file used by the elements dynamically, the elements resources are
updated.