fix(login-fire-form): email validation
Before this commit the email validation was quite too much simple and
was not letting pass some frequently used formats. Now, the validation
checks more precisely the email address format and allows more special
chars. It's alos possible to redefine the regex to apply on email
address when the form is confirmed.

close #175